feat: added rabbitmq support for integration events
-Refactor user ID handling to use UUID instead of int64 across the application

- Updated database schema to change user ID fields from BIGINT to UUID in users, user_bases, and military_operations tables.
- Modified SQL queries and generated code to reflect the new UUID type for user IDs.
- Adjusted mappers and repository methods to handle UUIDs for user IDs.
- Refactored token generation and validation to use UUID for user identification.
- Removed deprecated user registration and login endpoints, retaining only necessary user balance retrieval.
- Updated HTTP handlers and middleware to accommodate changes in user ID type.

+++ internal/game/infrastructure/db/mappers/military_operation_mapper.go
@@ -34,7 +34,7 @@ func MilitaryOperationFromDB(r gen.MilitaryOperation) *domain.MilitaryOperation 
 	op := &domain.MilitaryOperation{
 		ID:                int(r.ID),
 		Type:              domain.MilitaryOperationType(r.Type),
-		OwnerUserID:       int(r.OwnerUserID),
+		OwnerUserID:       r.OwnerUserID,
 		SourceBaseID:      int(r.SourceBaseID),
 		SourceCoordinates: domain.Vector2i{X: int(r.SourceX), Y: int(r.SourceY)},
 		TargetCoordinates: domain.Vector2i{X: int(r.TargetX), Y: int(r.TargetY)},
@@ -76,7 +76,7 @@ func InsertMilitaryOperationParamsFromDomain(op *domain.MilitaryOperation) gen.I
 
 	return gen.InsertMilitaryOperationParams{
 		Type:              string(op.Type),
-		OwnerUserID:       int64(op.OwnerUserID),
+		OwnerUserID:       op.OwnerUserID,
 		SourceBaseID:      int64(op.SourceBaseID),
 		SourceX:           int32(op.SourceCoordinates.X),
 		SourceY:           int32(op.SourceCoordinates.Y),
@@ -120,7 +120,7 @@ func UpdateMilitaryOperationParamsFromDomain(op *domain.MilitaryOperation) gen.U
 	return gen.UpdateMilitaryOperationParams{
 		ID:                int64(op.ID),
 		Type:              string(op.Type),
-		OwnerUserID:       int64(op.OwnerUserID),
+		OwnerUserID:       op.OwnerUserID,
 		SourceBaseID:      int64(op.SourceBaseID),
 		SourceX:           int32(op.SourceCoordinates.X),
 		SourceY:           int32(op.SourceCoordinates.Y),
